clean up payment-bot
